+++ URL-Ultimate-Filter-Surge.js
@@ -1,14 +1,14 @@
 /**
  * @file    URL-Ultimate-Filter-Surge.js
- * @version 46.56
- * @date    2026-08-20
+ * @version 46.57
+ * @date    2026-08-21
  * @rules   1535 total (312 domains, 439 critical paths, 401 path keywords, 109 param rules)
  * @build   SSOT Compiler — Dual-Target Compilation
  */
 
 const CONFIG = { DEBUG_MODE: false, AC_SCAN_MAX_LENGTH: 600 };
-const SCRIPT_VERSION = '46.56';
-const SCRIPT_BUILD = 'V46.56 (2026-08-20) | 1535 rules | 3232 tests';
+const SCRIPT_VERSION = '46.57';
+const SCRIPT_BUILD = 'V46.57 (2026-08-21) | 1535 rules | 3235 tests';
 const EMPTY_SET = new Set();
 
 const OAUTH_SAFE_HARBOR = {
@@ -1021,6 +1021,9 @@ const RULES = {
     PATH_EXEMPTIONS: new Map([
     ['id.atlassian.com', new Set([
         'RE:^/login(?:/|$)'
+      ])],
+    ['www.patreon.com', new Set([
+        'RE:^/api/tracking(?:/|$)'
       ])],
     ['storm.mg', new Set([
         '/_nuxt/track'